Use CTR mode (Crypto.Modes.ctr) instead of CBC.
First of all, we shouldn't be writing our own CBC code.  That
means that we should use Crypto.Padding.

Now, CTR mode is widely regarded as secure (as well as CBC),
however it doesn't need any padding.  Using CTR let us shave 1 to
16 bytes in the encrypted message, which translates to 1 to 22
less bytes in the final cookie.  That's why we've switched to CTR mode.