Use cprng-aes instead of our own implementation of AES as a CPRNG.
This adds a new dependency, but I think it's for the better.  It
doesn't remove a lot of code, but we get some more features, such
starting with a random counter value and having an IV.
Performance is 8% worse, but since I've also bumped the threshold
to 100.000, it should be better =).

    An IV used to generate IVs?
    IVception!